chore(pipeline): preserve legacy cli.js, slim down active cli.js, backup as cli.legacy.js
